Move strategy package
