+++ elastic-job-lite-core/src/main/java/org/apache/shardingsphere/elasticjob/lite/handler/sharding/JobShardingStrategyFactory.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.api.strategy;
+package org.apache.shardingsphere.elasticjob.lite.handler.sharding;
 
 import com.google.common.base.Strings;
 import lombok.AccessLevel;
@@ -27,7 +27,7 @@ import java.util.Map;
 import java.util.ServiceLoader;
 
 /**
- * Job sharding strategy factory.
+ * Job sharding sharding factory.
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class JobShardingStrategyFactory {
@@ -43,17 +43,17 @@ public final class JobShardingStrategyFactory {
     }
     
     /**
-     * Get job sharding strategy.
+     * Get job sharding sharding.
      * 
-     * @param type job sharding strategy type
-     * @return job sharding strategy
+     * @param type job sharding sharding type
+     * @return job sharding sharding
      */
     public static JobShardingStrategy getStrategy(final String type) {
         if (Strings.isNullOrEmpty(type)) {
             return STRATEGIES.get(DEFAULT_STRATEGY);
         }
         if (!STRATEGIES.containsKey(type)) {
-            throw new JobConfigurationException("Can not find sharding strategy type '%s'.", type);
+            throw new JobConfigurationException("Can not find sharding sharding type '%s'.", type);
         }
         return STRATEGIES.get(type);
     }
