+++ elastic-job-lite-core/src/main/java/org/apache/shardingsphere/elasticjob/lite/handler/sharding/impl/AverageAllocationJobShardingStrategy.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,10 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.api.strategy.impl;
+package org.apache.shardingsphere.elasticjob.lite.handler.sharding.impl;
 
-import org.apache.shardingsphere.elasticjob.lite.api.strategy.JobInstance;
-import org.apache.shardingsphere.elasticjob.lite.api.strategy.JobShardingStrategy;
+import org.apache.shardingsphere.elasticjob.lite.handler.sharding.JobInstance;
+import org.apache.shardingsphere.elasticjob.lite.handler.sharding.JobShardingStrategy;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -27,7 +27,7 @@ import java.util.List;
 import java.util.Map;
 
 /**
- * Sharding strategy which for average by sharding item.
+ * Sharding sharding which for average by sharding item.
  * 
  * <p>
  * If the job server number and sharding count cannot be divided,
