+++ elastic-job-lite-core/src/main/resources/META-INF/services/org.apache.shardingsphere.elasticjob.lite.handler.sharding.JobShardingStrategy
@@ -15,6 +15,6 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.elasticjob.lite.api.strategy.impl.AverageAllocationJobShardingStrategy
-org.apache.shardingsphere.elasticjob.lite.api.strategy.impl.OdevitySortByNameJobShardingStrategy
-org.apache.shardingsphere.elasticjob.lite.api.strategy.impl.RoundRobinByNameJobShardingStrategy
+org.apache.shardingsphere.elasticjob.lite.handler.sharding.impl.AverageAllocationJobShardingStrategy
+org.apache.shardingsphere.elasticjob.lite.handler.sharding.impl.OdevitySortByNameJobShardingStrategy
+org.apache.shardingsphere.elasticjob.lite.handler.sharding.impl.RoundRobinByNameJobShardingStrategy
