+++ elastic-job-lite-core/src/test/java/org/apache/shardingsphere/elasticjob/lite/handler/sharding/impl/OdevitySortByNameJobShardingStrategyTest.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,9 +15,9 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.api.strategy.impl;
+package org.apache.shardingsphere.elasticjob.lite.handler.sharding.impl;
 
-import org.apache.shardingsphere.elasticjob.lite.api.strategy.JobInstance;
+import org.apache.shardingsphere.elasticjob.lite.handler.sharding.JobInstance;
 import org.junit.Test;
 
 import java.util.Arrays;
